Refactor HTML/JS to create locale-specific pages.
For now, just have es-mx locale.
