+++ site/css/index.css
@@ -1,0 +1,13 @@
+main {
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    grid-template-rows: 1fr;
+    grid-column-gap: 10px;
+    grid-row-gap: 10px;
+    justify-items: center;
+    align-items: center;
+}
+
+main a { grid-area: 1 / 1 / 2 / 2; }
+main img { grid-area: 1 / 2 / 2 / 3; }
+
