--- site/debug.html
@@ -1,53 +1,0 @@
-<!DOCTYPE html>
-<!---
-// Copyright (c) 2019 Eamonn O'Brien-Strain All rights reserved. This
-// program and the accompanying materials are made available under the
-// terms of the Eclipse Public License v1.0 which accompanies this
-// distribution, and is available at
-// http://www.eclipse.org/legal/epl-v10.html
------>
-<html>
-
-<head>
-    <!-- Global site tag (gtag.js) - Google Analytics -->
-    <script async src="https://www.googletagmanager.com/gtag/js?id=UA-6341708-3"></script>
-    <script>
-        window.dataLayer = window.dataLayer || [];
-        function gtag() { dataLayer.push(arguments); }
-        gtag('js', new Date());
-        gtag('config', 'UA-6341708-3');
-    </script>
-    <meta charset="utf-8">
-    <meta http-equiv="x-ua-compatible" content="ie=edge">
-    <title>Kartoj</title>
-    <meta name="description" content="Immersive Image-Based Language Learning">
-    <meta name="author" content="@eob">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link href="//fonts.googleapis.com/css?family=Roboto&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="css/debug.css">
-    <link rel="icon" type="image/png" href="/images/noun_Man_1912182_32x32.png">
-    <script src="debug_compiled.js" defer></script>
-</head>
-
-<body>
-    <header>
-        <h1><a href="/"><img src="/images/noun_Man_1912182_32x32.png" /></a>Kartoj &mdash; Debug</h1>
-    </header>
-    <table id="table">
-        <tr>
-            <th>Score</th>
-            <th>Responses</th>
-            <th>Query</th>
-            <th>Images</th>
-        </tr>
-    </table>
-    <footer>
-        <a href="credit.html">Kredito</a>
-        -
-        <a href="privacy.html">Privateco</a>
-        -
-        <a href="deck.html">Superrigardo</a>
-    </footer>
-</body>
-
-</html>
