+++ site/src/deck_es_mx.js
@@ -1,0 +1,12 @@
+// Copyright (c) 2019 Eamonn O'Brien-Strain All rights reserved. This
+// program and the accompanying materials are made available under the
+// terms of the Eclipse Public License v1.0 which accompanies this
+// distribution, and is available at
+// http://www.eclipse.org/legal/epl-v10.html
+
+// @ts-check
+
+import { mergiWords } from './words_es_mx.js'
+import deck from './deck.js'
+
+deck('es', 'mx', mergiWords)
