+++ template/card_bundled.json
