+++ template/debug.html
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<!---
+// Copyright (c) 2019 Eamonn O'Brien-Strain All rights reserved. This
+// program and the accompanying materials are made available under the
+// terms of the Eclipse Public License v1.0 which accompanies this
+// distribution, and is available at
+// http://www.eclipse.org/legal/epl-v10.html
+----->
+<html>
+{{> head}}
+
+<body>
+    {{> header}}
+    <table id="table">
+        <tr>
+            <th>Score</th>
+            <th>Responses</th>
+            <th>Query</th>
+            <th>Images</th>
+        </tr>
+    </table>
+    {{> footer}}
+</body>
+
+</html>
