+++ template/deck_es_mx.json
@@ -3,6 +3,8 @@
     "modulepreloads": [],
     "stylesheets": ["app", "deck"],
     "scripts": [
-        {"src": "deck_compiled.js", "type": "defer"}
-    ]
+        {"src": "deck_es_mx_compiled.js", "type": "defer"}
+    ],
+    "lang": "es",
+    "country": "mx"
 }
