+++ template/index.html
@@ -3,34 +3,12 @@
 {{> head}}
 
 <body>
-    <main id="card" class="initial">
-        <section id="front" class="unflpped">
-            <nav />
-        </section>
-        <section id="back" class="unflpped">
-            <nav>
-                <!-- Icon attribution: wrong by Adrien Coquet from the Noun Project https://thenounproject.com/coquet_adrien/collection/head-mind-thought-creative/?i=1912187 -->
-                <svg id="wrong" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 100 125">
-                    <path
-                        d="M90.8 55.5l-8.1-12.4c-.4-.7-.6-1.6-.7-2.4-1.4-16.2-15-29-31.5-29.1 1.6 3.2 2.4 6.8 2.4 10.6 0 13.5-11 24.4-24.4 24.4-3.6 0-7-.8-10-2.2.3 10.1 5.3 19.9 12.8 26.3.9.8 1.5 1.9 1.5 3.1v18.3c0 1.5 1.1 2.6 2.5 2.8l27.4 2.4c1.6.1 3-1.1 3-2.8V81.8c0-.8.7-1.5 1.5-1.5h9.6c3 0 5.4-2.4 5.4-5.4V62.1c0-.6.5-1 1-1h4.7c2.8-.1 4.5-3.2 2.9-5.6z" />
-                    <path
-                        d="M48.2 22.3c0-10.9-8.9-19.8-19.8-19.8S8.6 11.4 8.6 22.3s8.9 19.8 19.8 19.8 19.8-8.9 19.8-19.8zm-25.8 8.2c-.6 0-1.2-.2-1.6-.7-.9-.9-.9-2.3 0-3.2l4.4-4.4-4.4-4.4c-.9-.9-.9-2.3 0-3.2.9-.9 2.3-.9 3.2 0l4.4 4.4 4.4-4.4c.9-.9 2.3-.9 3.2 0 .9.9.9 2.3 0 3.2l-4.4 4.4 4.4 4.4c.9.9.9 2.3 0 3.2-.4.4-1 .7-1.6.7-.6 0-1.2-.2-1.6-.7l-4.4-4.4-4.4 4.5c-.4.4-1 .6-1.6.6z" />
-                </svg>
-                <!-- Icon attribution: Agree by Adrien Coquet from the Noun Project https://thenounproject.com/term/agree/1912218/ -->
-                <svg id="meh" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 100 125">
-                    <path
-                        d="M90.8 55.5l-8.1-12.4c-.4-.7-.6-1.6-.7-2.4-1.4-16.2-15-29-31.6-29.1 1.6 3.2 2.4 6.8 2.4 10.6 0 13.5-11 24.4-24.4 24.4-3.6 0-7-.8-10-2.2.3 10.1 5.3 19.9 12.8 26.3.9.8 1.5 1.9 1.5 3.1v18.3c0 1.5 1.1 2.6 2.5 2.8l27.4 2.4c1.6.1 3-1.1 3-2.8V81.8c0-.8.7-1.5 1.5-1.5h9.6c2.9 0 5.3-2.4 5.3-5.4V62.1c0-.6.5-1 1-1h4.7c3-.1 4.6-3.2 3.1-5.6z" />
-                    <path
-                        d="M48.2 22.3c0-10.9-8.9-19.8-19.8-19.8S8.6 11.4 8.6 22.3s8.9 19.8 19.8 19.8 19.8-8.9 19.8-19.8zm-21.6 7.2s-.1 0 0 0c-.7 0-1.2-.2-1.7-.7l-5.1-5.1c-.9-.9-.9-2.3 0-3.2.9-.9 2.3-.9 3.2 0l3.4 3.4 6.7-7.6c.8-.9 2.3-1 3.2-.2.9.8 1 2.3.2 3.2l-8.3 9.4c-.4.5-1 .8-1.6.8z" />
-                </svg>
-                <svg id="correct" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 100 125">
-                    <path
-                        d="M90.8 55.5l-8.1-12.4c-.4-.7-.6-1.6-.7-2.4-1.4-16.2-15-29-31.6-29.1 1.6 3.2 2.4 6.8 2.4 10.6 0 13.5-11 24.4-24.4 24.4-3.6 0-7-.8-10-2.2.3 10.1 5.3 19.9 12.8 26.3.9.8 1.5 1.9 1.5 3.1v18.3c0 1.5 1.1 2.6 2.5 2.8l27.4 2.4c1.6.1 3-1.1 3-2.8V81.8c0-.8.7-1.5 1.5-1.5h9.6c2.9 0 5.3-2.4 5.3-5.4V62.1c0-.6.5-1 1-1h4.7c3-.1 4.6-3.2 3.1-5.6z" />
-                    <path
-                        d="M48.2 22.3c0-10.9-8.9-19.8-19.8-19.8S8.6 11.4 8.6 22.3s8.9 19.8 19.8 19.8 19.8-8.9 19.8-19.8zm-21.6 7.2s-.1 0 0 0c-.7 0-1.2-.2-1.7-.7l-5.1-5.1c-.9-.9-.9-2.3 0-3.2.9-.9 2.3-.9 3.2 0l3.4 3.4 6.7-7.6c.8-.9 2.3-1 3.2-.2.9.8 1 2.3.2 3.2l-8.3 9.4c-.4.5-1 .8-1.6.8z" />
-                </svg>
-            </nav>
-        </section>
+    {{> header}}
+    <main>
+        {{#locales}}
+        <img width="100" height="57" src="images/flag_{{country}}.webp">
+        <a href="card_{{lang}}_{{country}}.html">{{langName}}</a>
+        {{/locales}}
     </main>
     {{> footer}}
 </body>
